wardrobe_ml_system/requirements.txt
# Core ML
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Web Framework
flask>=2.0.0

# Database
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=0.19.0
